alura-django-rest-aluraflix/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1
drf-yasg==1.21.3
isort
black
